Change token genration to tirst_time_guardian_person file and add token to databse
