+++ babyOne/app/src/main/java/com/example/babyone/FirstTimeGuardian.java
@@ -127,6 +127,7 @@ public class FirstTimeGuardian extends AppCompatActivity {
 
         String collectionName = "guardians";
         FirestoreHelper.addToFirestore(collectionName, registrationInfo, this, FirstTimeGuardian.this);
+        startActivity(new Intent(FirstTimeGuardian.this, ProfileActivity.class).setFlags(Intent.FLAG_ACTIVITY_NEW_TASK));
 
     }
 }
