+++ babyOne/app/src/main/java/com/example/babyone/MainActivity.java
@@ -115,23 +115,23 @@ public class MainActivity extends AppCompatActivity {
                                     }
 
                                     // Notification Handle
-                                    FirebaseMessaging.getInstance().getToken()
-                                            .addOnCompleteListener(new OnCompleteListener<String>() {
-                                                @Override
-                                                public void onComplete(@NonNull Task<String> task) {
-                                                    if (!task.isSuccessful()) {
-                                                        Log.w(TAG, "Fetching FCM registration token failed", task.getException());
-                                                        return;
-                                                    }
-
-                                                    // Get new FCM registration token
-                                                    String token = task.getResult();
-
-                                                    // Log and toast
-                                                    Log.d(TAG, "Token: "+token);
-                                                    Toast.makeText(MainActivity.this, token, Toast.LENGTH_SHORT).show();
-                                                }
-                                            });
+//                                    FirebaseMessaging.getInstance().getToken()
+//                                            .addOnCompleteListener(new OnCompleteListener<String>() {
+//                                                @Override
+//                                                public void onComplete(@NonNull Task<String> task) {
+//                                                    if (!task.isSuccessful()) {
+//                                                        Log.w(TAG, "Fetching FCM registration token failed", task.getException());
+//                                                        return;
+//                                                    }
+//
+//                                                    // Get new FCM registration token
+//                                                    String token = task.getResult();
+//
+//                                                    // Log and toast
+//                                                    Log.d(TAG, "Token: "+token);
+//                                                    Toast.makeText(MainActivity.this, token, Toast.LENGTH_SHORT).show();
+//                                                }
+//                                            });
 
                                     /*// When task is successful redirect to profile activity display Toast
                                     startActivity(new Intent(MainActivity.this, ProfileActivity.class).setFlags(Intent.FLAG_ACTIVITY_NEW_TASK));
