+++ babyOne/app/src/main/java/com/example/babyone/first_time_guardian_personal.java
@@ -1,14 +1,23 @@
 package com.example.babyone;
 
+import static android.content.ContentValues.TAG;
+
 import android.os.Bundle;
 
+import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.EditText;
 import android.widget.RadioGroup;
+import android.widget.Toast;
+
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.messaging.FirebaseMessaging;
 
 import java.util.HashMap;
 
@@ -30,6 +39,24 @@ public class first_time_guardian_personal extends Fragment {
         radgroupEPRelation = view.findViewById(R.id.radgroupEPRelation);
         //TODO: Implement relationship to the baby radio button
 
+        // Create a token and add it to database
+        FirebaseMessaging.getInstance().getToken()
+                .addOnCompleteListener(new OnCompleteListener<String>() {
+                    @Override
+                    public void onComplete(@NonNull Task<String> task) {
+                        if (!task.isSuccessful()) {
+                            Log.w(TAG, "Fetching FCM registration token failed", task.getException());
+                            return;
+                        }
+
+                        // Get new FCM registration token
+                        String token = task.getResult();
+
+                        // Log and toast
+                        Log.d(TAG, "Token: "+token);
+                        editTextMap.put("token",token);
+                    }
+                });
 
         // Inflate the layout for this fragment
         return view;
@@ -66,6 +93,7 @@ public class first_time_guardian_personal extends Fragment {
         if (!relationship.isEmpty()) {
             editTextMap.put("relationship", relationship);
         }
+
         return editTextMap;
     }
 }
